MIPS Project #1
You will write a MIPS program that reads a string of up to 8 characters from user input. If
the string has only the characters from '0' to '9' and from 'a' to 'f' and from 'A' to 'F', the
program prints out the unsigned decimal integer corresponding to the hexadecimal number
in the string. Otherwise, the program prints out the message of "Invalid hexadecimal
number.".